requirements.txt
discord.py>=2.4.0
psutil>=5.9.0
python-dotenv>=1.0.0